+++ application/src/main/java/common/bolts/transactional/ob/OBBolt.java
@@ -38,7 +38,7 @@ public abstract class OBBolt extends TransactionalBolt {
      * @param bid
      * @param timestamp
      */
-    protected void auth(long bid, Long timestamp) {
+    protected void auth(double bid, Long timestamp) {
 //        System.out.println(generatedString);
 //        stateless_task.random_compute(5);
     }
@@ -156,7 +156,7 @@ public abstract class OBBolt extends TransactionalBolt {
             collector.emit(event.getBid(), event.biding_result, event.getTimestamp());//the tuple is finished finally.
         } else {
             if (enable_latency_measurement) {
-                sink.execute(new Tuple(event.getBid(), this.thread_Id, context, new GeneralMsg<>(DEFAULT_STREAM_ID, event.biding_result, event.getTimestamp())));//(long bid, int sourceId, TopologyContext context, Message message)
+                sink.execute(new Tuple(event.getBid(), this.thread_Id, context, new GeneralMsg<>(DEFAULT_STREAM_ID, event.biding_result, event.getTimestamp())));//(double bid, int sourceId, TopologyContext context, Message message)
             }
         }
     }
@@ -166,7 +166,7 @@ public abstract class OBBolt extends TransactionalBolt {
             collector.emit(event.getBid(), event.alert_result, event.getTimestamp());//the tuple is finished finally.
         } else {
             if (enable_latency_measurement) {
-                sink.execute(new Tuple(event.getBid(), this.thread_Id, context, new GeneralMsg<>(DEFAULT_STREAM_ID, event.alert_result, event.getTimestamp())));//(long bid, int sourceId, TopologyContext context, Message message)
+                sink.execute(new Tuple(event.getBid(), this.thread_Id, context, new GeneralMsg<>(DEFAULT_STREAM_ID, event.alert_result, event.getTimestamp())));//(double bid, int sourceId, TopologyContext context, Message message)
             }
         }
     }
@@ -176,7 +176,7 @@ public abstract class OBBolt extends TransactionalBolt {
             collector.emit(event.getBid(), event.topping_result, event.getTimestamp());//the tuple is finished finally.
         } else {
             if (enable_latency_measurement) {
-                sink.execute(new Tuple(event.getBid(), this.thread_Id, context, new GeneralMsg<>(DEFAULT_STREAM_ID, event.topping_result, event.getTimestamp())));//(long bid, int sourceId, TopologyContext context, Message message)
+                sink.execute(new Tuple(event.getBid(), this.thread_Id, context, new GeneralMsg<>(DEFAULT_STREAM_ID, event.topping_result, event.getTimestamp())));//(double bid, int sourceId, TopologyContext context, Message message)
             }
         }
     }
